// src/types/schemas/series.ts
/* eslint-disable */
/**
 * This file was automatically generated by json-schema-to-typescript.
 * DO NOT MODIFY IT BY HAND. Instead, modify the source JSONSchema file,
 * and run json-schema-to-typescript to regenerate this file.
 */

export interface SeriesSchema {
  name: string;
  genre: "Open-Wheel" | "Endurance" | "GT / Touring" | "Stock Car" | "Rally" | "Off-Road" | "Motorcycle" | "Other";
  seriesUrl: string;
  logoUrl: string;
  events: EventSchema[];
}
export interface EventSchema {
  eventUrl: string;
  eventStartDate: string;
  eventEndDate: string;
  eventName: string;
  location: string;
  sessions: SessionSchema[];
}
export interface SessionSchema {
  sessionDatetime: string;
  sessionName: string;
  sessionType: "Practice" | "Qualifying" | "Race" | "Testing" | "Other";
}